Add an agent config option to diable per tenancy usage metrics.
