+++ agent/config/testdata/TestRuntimeConfig_Sanitize.golden
@@ -208,11 +208,11 @@
     "ExposeMaxPort": 0,
     "ExposeMinPort": 0,
     "GRPCAddrs": [],
+    "GRPCKeepaliveInterval": "0s",
+    "GRPCKeepaliveTimeout": "0s",
     "GRPCPort": 0,
     "GRPCTLSAddrs": [],
     "GRPCTLSPort": 0,
-    "GRPCKeepaliveInterval": "0s",
-    "GRPCKeepaliveTimeout": "0s",
     "GossipLANGossipInterval": "0s",
     "GossipLANGossipNodes": 0,
     "GossipLANProbeInterval": "0s",
@@ -472,6 +472,7 @@
         "CirconusSubmissionURL": "",
         "Disable": false,
         "DisableHostname": false,
+        "DisablePerTenancyUsageMetrics": false,
         "DogstatsdAddr": "",
         "DogstatsdTags": [],
         "EnableHostMetrics": false,
