+++ agent/config/testdata/full-config.json
@@ -841,7 +841,8 @@
     "metrics_prefix": "ftO6DySn",
     "prometheus_retention_time": "15s",
     "statsd_address": "drce87cy",
-    "statsite_address": "HpFwKB8R"
+    "statsite_address": "HpFwKB8R",
+    "disable_per_tenancy_usage_metrics": true
   },
   "tls": {
     "defaults": {
