+++ website/content/docs/agent/config/config-files.mdx
@@ -1842,6 +1842,9 @@ subsystem that provides Consul's service mesh capabilities.
   - `disable_hostname` ((#telemetry-disable_hostname))
     Set to `true` to stop prepending the machine's hostname to gauge-type metrics. Default is `false`.
 
+  - `disable_per_tenancy_usage_metrics` ((#telemetry-disable_per_tenancy_usage_metrics))
+    Set to `true` to exclude tenancy labels from usage metrics. This significantly decreases CPU utilization in clusters with many admin partitions or namespaces.
+
   - `dogstatsd_addr` ((#telemetry-dogstatsd_addr)) This provides the address
     of a DogStatsD instance in the format `host:port`. DogStatsD is a protocol-compatible
     flavor of statsd, with the added ability to decorate metrics with tags and event
